\name{ETOPO5}
\alias{ETOPO5}

\docType{data}
\title{Etopo5 topographic data}
\description{
Etopo5 topographic data
}
\usage{data(ETOPO5)}
\format{
 matrix of world elevation from 0-360 degrees long
}
\details{
  See web site:
   http://www.ngdc.noaa.gov/mgg/global/etopo5.HTML
}

\references{


Data Announcement 88-MGG-02, Digital relief of the Surface of the Earth. NOAA, National Geophysical Data Center, Boulder, Colorado, 1988.

  
}
\examples{
data(ETOPO5)
## maybe str(ETOPO5) ; plot(ETOPO5) ...


}
\keyword{datasets}
