\name{geomapdata-package}
\alias{geomapdata-package}
\alias{geomapdata}
\docType{package}
\title{
geomapdata
}
\description{
Topographic and Geologic Mapping
}
\details{
\tabular{ll}{
Package: \tab geomapdata\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2007-10-23\cr
License: \tab GPL\cr
}
Set of data for making Maps, Topographic Maps, Perspective plots, geological databases.
}
\author{
Jonathan M. Lees<jonathan.lees.edu>
Maintainer:Jonathan M. Lees<jonathan.lees@unc.edu>
}
\references{

Lees, J. M., Geotouch: Software for Three and Four Dimensional GIS in the Earth Sciences, Computers & Geosciences, 26, 7, 751-761, 2000.

}

\keyword{ package }
\seealso{GEOmap}
\examples{

data(worldmap)


}
