\name{usacity}
\alias{usacity}
\alias{worldcity}

\docType{data}
\title{City Locations and Populations(USA)}
\description{
  point data set showing cities locations
  and populations.
}
\usage{data(usacity)}
\format{
  \item{name}{name of city}
  \item{lat}{latitude}
  \item{lon}{longitude}
  \item{p}{population} 
}
\details{
World cities have no population (yet).
}

\examples{
data(usacity)
## maybe str(usacity) ; plot(usacity) ...
}
\keyword{datasets}
